requirements.txt
Django==4.0.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Jinja2==3.0.3
whitenoise==6.3.0
pandas
matplotlib